update hash stuff
