+++ package.json
@@ -5,14 +5,14 @@
     "scripts": {
         "tauri": "tauri",
         "dev": "vite dev",
-        "build": "prebuild-commit-hash && vite build",
+        "build": "npm run write-commit-hash && vite build",
         "preview": "vite preview --host 0.0.0.0 --port 80",
         "check": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json",
         "check:watch": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json --watch",
         "format": "prettier --write 'src/**/*.{css,html,js,svelte,ts}'",
         "lint": "npm run format && npm run check",
-        "prebuild-commit-hash": "node scripts/get-commit-hash.js"
-      },
+        "write-commit-hash": "node ./scripts/write-commit-hash.cjs"
+    },
     "devDependencies": {
         "@castlenine/vite-remove-attribute": "^1.0.2",
         "@sveltejs/adapter-auto": "^3.0.0",
