+++ src/commit-hash.ts
@@ -1,0 +1,1 @@
+export const COMMIT_HASH = '746abadd15509759f65d14d74fa24eccb14c6199';
