+++ svelte.config.js
@@ -1,36 +1,26 @@
 import { vitePreprocess } from "@sveltejs/vite-plugin-svelte"
 import adapter from '@sveltejs/adapter-static'
 import path from "node:path"
-import { execSync } from 'child_process'
 
-// Function to get the latest Git commit hash
-function getCommitHash() {
-    try {
-        return execSync('git rev-parse HEAD').toString().trim()
-    } catch (error) {
-        console.error('Error fetching commit hash:', error)
-        return 'unknown'
-    }
-}
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-    preprocess: [
-        vitePreprocess({}),
-    ],
-    kit: {
-        adapter: adapter(),
-        vite: {
-            define: {
-                __COMMIT_HASH__: JSON.stringify(getCommitHash()),
-            },
-            resolve: {
-                alias: {
-                    $lib: path.resolve('src/lib')
-                }
-            }
-        }
-    }
+  preprocess: [
+    vitePreprocess({}),
+  ],
+  kit: {
+    // adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
+    // If your environment is not supported or you settled on a specific environment, switch out the adapter.
+    // See https://kit.svelte.dev/docs/adapters for more information about adapters.
+    adapter: adapter(),
+    // vite: {
+    //   resolve: {
+    //     alias: {
+    //       $lib: path.resolve('src/lib')
+    //     }
+    //   }
+    // }
+  },
 }
 
 export default config
